requirements.txt
# horovod==0.22
tensorflow-gpu<2 #1.15
tensorflow_probability<0.9
